refactor: add firecracker host runtime leases
Amp-Thread-ID: https://ampcode.com/threads/T-019d9e99-73fd-770e-a7d3-9693f950ce64
